Move object related functions to util
